+++ docker.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-sudo docker build -t derailed-api .
-sudo docker stop derailed-api
-sudo docker rm -v derailed-api
-sudo docker run --name derailed-api -p 5000:5000 -d derailed-api
+sudo docker build -t twattle-api .
+sudo docker stop twattle-api
+sudo docker rm -v twattle-api
+sudo docker run --name twattle-api -p 5000:5000 -d twattle-api
